wip: move the "check_threshold" util to new module files
