+++ circuit/src/threshold/mod.rs
@@ -1,0 +1,2 @@
+/// Native version of checking score threshold
+pub mod native;
